refactor: move folder image generation to service
